refactor: move file handling tasks to this package
file handling is found to be implemented in every package which
goes against the idea of separation of concerns. to keep things clean,
only fileio will handle the opening and closing of the file from now.
other packages will just operate on the buffer. it should be noted that
the current implementation will mess up if the file being read is too
big.